+++ loess/src/lib.rs
@@ -99,7 +99,7 @@ impl IntoTokens for Error {
 
 		#[allow(unused_variables, unused_mut)] // Not suppressed because it's the same crate.
 		{
-			quote_into_same_site! (span, root, tokens, [
+			quote_into_with_exact_span! (span, root, tokens, [
 				{#error {#paste message}};
 			]);
 		}