+++ loess/tests/quote_expansions.rs
@@ -1,8 +1,8 @@
 #![deny(unused_variables)] // At least for now, this is used to detect missing expansions.
 
 use loess::{
-	quote_into_call_site, quote_into_mixed_site, quote_into_same_site, raw_quote_into_call_site,
-	raw_quote_into_mixed_site, raw_quote_into_same_site,
+	quote_into_call_site, quote_into_mixed_site, quote_into_with_exact_span,
+	raw_quote_into_call_site, raw_quote_into_mixed_site, raw_quote_into_with_exact_span,
 };
 use proc_macro2::{Span, TokenStream, TokenTree};
 
@@ -18,8 +18,6 @@ macro_rules! test {
 	}};
 }
 
-//TODO: Wrap the six functions below into tests.
-
 #[test]
 pub fn mixed_site() {
 	test!(let (span, root, tokens), quote_into_mixed_site!(span, root, tokens, [....... .....]), "....... .....");
@@ -27,7 +25,7 @@ pub fn mixed_site() {
 
 #[test]
 pub fn same_site() {
-	test!(let (span, root, tokens), quote_into_same_site!(span, root, tokens, [....... .....]), "....... .....");
+	test!(let (span, root, tokens), quote_into_with_exact_span!(span, root, tokens, [....... .....]), "....... .....");
 }
 
 #[test]
@@ -42,7 +40,7 @@ pub fn mixed_site_raw() {
 
 #[test]
 pub fn same_site_raw() {
-	test!(let (span, _root, tokens), raw_quote_into_same_site!(span, tokens, [....... .....]), "....... .....");
+	test!(let (span, _root, tokens), raw_quote_into_with_exact_span!(span, tokens, [....... .....]), "....... .....");
 }
 
 #[test]
@@ -58,7 +56,7 @@ pub fn long_punctuation() {
 #[test]
 pub fn error() {
 	let mut custom_root = TokenStream::new();
-	raw_quote_into_same_site!(Span::call_site(), &mut custom_root, [::custom::root]);
+	raw_quote_into_with_exact_span!(Span::call_site(), &mut custom_root, [::custom::root]);
 	test!(
 		let (span, _root, tokens),
 		quote_into_mixed_site!(span, &custom_root, tokens, [{#error "This is an error message."}]),
@@ -69,11 +67,24 @@ pub fn error() {
 #[test]
 pub fn root() {
 	let mut custom_root = TokenStream::new();
-	raw_quote_into_same_site!(Span::call_site(), &mut custom_root, [::custom::root]);
+	raw_quote_into_with_exact_span!(Span::call_site(), &mut custom_root, [::custom::root]);
 	test!(let (span, _root, tokens), quote_into_mixed_site!(span, &custom_root, tokens, [{#root}]), ":: custom :: root");
 }
 
-//TODO: More tests.
+#[test]
+pub fn let_and_span_directives() {
+	test!(let (span, root, tokens), quote_into_mixed_site!(span, root, tokens, [
+		{#let a = Span::mixed_site();}
+		{#let b = Span::call_site()}
+		{#let Some(c) = Some(Span::mixed_site()), else { unreachable!() };}
+		{#let Some(_d) = Some(Span::call_site()), else { unreachable!() }}
+		{#mixed_site mx }
+		{#call_site cs }
+		{#located_at a, a_ }
+		{#resolved_at b, b_ }
+		{#with_exact_span c, c_ }
+	]), "mx cs a_ b_ c_");
+}
 
 #[test]
 pub fn r#return() {
@@ -297,7 +308,159 @@ pub fn break_from_for_with_label() {
 	test!(let (span, root, tokens), break_from_for_with_label(span, root, tokens), "once");
 }
 
-//TODO: More tests!
+#[test]
+pub fn continue_in_for() {
+	fn continue_in_for(span: Span, root: &TokenStream, tokens: &mut impl Extend<TokenTree>) {
+		quote_into_mixed_site!(span, root, tokens, [
+			{#for _ in 0..2,
+				twice
+				{#if true, {#continue;}}
+				never
+			}
+		]);
+	}
+
+	test!(let (span, root, tokens), continue_in_for(span, root, tokens), "twice twice");
+}
+
+#[test]
+pub fn continue_in_for_with_label() {
+	fn continue_in_for_with_label(
+		span: Span,
+		root: &TokenStream,
+		tokens: &mut impl Extend<TokenTree>,
+	) {
+		quote_into_mixed_site!(span, root, tokens, [
+			{#'my_label: for _ in 0..2,
+				twice
+				{#if true, {#continue 'my_label;}}
+				never
+			}
+		]);
+	}
+
+	test!(let (span, root, tokens), continue_in_for_with_label(span, root, tokens), "twice twice");
+}
+
+#[test]
+pub fn for_else() {
+	fn for_else(span: Span, root: &TokenStream, tokens: &mut impl Extend<TokenTree>) {
+		quote_into_mixed_site!(span, root, tokens, [
+			{#for _ in 0..0,
+				never
+			} {#else,
+				once
+			}
+		]);
+	}
+
+	test!(let (span, root, tokens), for_else(span, root, tokens), "once");
+}
+
+#[test]
+pub fn for_not_else() {
+	fn for_not_else(span: Span, root: &TokenStream, tokens: &mut impl Extend<TokenTree>) {
+		quote_into_mixed_site!(span, root, tokens, [
+			{#if false, }
+			{#for _ in 0..2,
+				twice
+			} {#else,
+				never
+			}
+		]);
+	}
+
+	test!(let (span, root, tokens), for_not_else(span, root, tokens), "twice twice");
+}
+
+#[test]
+pub fn while_continue_with_label() {
+	fn while_continue_with_label(
+		span: Span,
+		root: &TokenStream,
+		tokens: &mut impl Extend<TokenTree>,
+	) {
+		let mut i = 0;
+		quote_into_mixed_site!(span, root, tokens, [
+			{#'my_label: while i < 2,
+				twice
+				{#let _ = i += 1;} // Not recommended, obviously.
+				{#if true, {#continue 'my_label;}}
+				never
+			}
+		]);
+	}
+
+	test!(let (span, root, tokens), while_continue_with_label(span, root, tokens), "twice twice");
+}
+
+#[test]
+pub fn while_else() {
+	fn while_else(span: Span, root: &TokenStream, tokens: &mut impl Extend<TokenTree>) {
+		quote_into_mixed_site!(span, root, tokens, [
+			{#while false,
+				never
+			} {#else,
+				once
+			}
+		]);
+	}
+
+	test!(let (span, root, tokens), while_else(span, root, tokens), "once");
+}
+
+#[test]
+pub fn while_not_else() {
+	fn while_not_else(span: Span, root: &TokenStream, tokens: &mut impl Extend<TokenTree>) {
+		quote_into_mixed_site!(span, root, tokens, [
+			{#if false, }
+			{#while true,
+				once
+				{#break;}
+			} {#else,
+				never
+			}
+		]);
+	}
+
+	test!(let (span, root, tokens), while_not_else(span, root, tokens), "once");
+}
+
+#[test]
+pub fn while_let() {
+	fn while_let(span: Span, root: &TokenStream, tokens: &mut impl Extend<TokenTree>) {
+		let mut condition = Some(true);
+		quote_into_mixed_site!(span, root, tokens, [
+			{#while let Some(_) = condition,
+				once
+				{#let _ = condition = None;}
+			}
+		]);
+	}
+
+	test!(let (span, root, tokens), while_let(span, root, tokens), "once");
+}
+
+#[test]
+pub fn else_scoping() {
+	// This checks that the "calling convention" of `enter_else` is correct.
+	fn else_scoping(span: Span, root: &TokenStream, tokens: &mut impl Extend<TokenTree>) {
+		quote_into_mixed_site!(span, root, tokens, [
+			{#if false, never }
+
+			// It's generally not great to have tokens between `#if` and `#else` like this,
+			// but online directives on the same level, they do not reset the fallback flag.
+			{{#else, never }}
+			[{#else, never }]
+			({#else, never })
+
+			{#else, once }
+			{#else, never }
+		]);
+	}
+
+	test!(let (span, root, tokens), else_scoping(span, root, tokens), "{ } [] () once");
+}
 
 #[test]
 fn braced() {